backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25

# Audio Processing
librosa==0.10.1
mutagen==1.47.0
soundfile==0.12.1
audioread==3.0.1

# Machine Learning
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.37.2
sentence-transformers==2.3.1

# AI/LLM
anthropic>=0.79.0
openai>=1.0.0

# External APIs
pylast==5.2.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
numpy==1.26.3
pandas==2.2.0
scipy>=1.2.0,<1.12.0  # Pin for librosa 0.10.1 compatibility

# HTTP Client
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Logging
python-json-logger==2.0.7

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
